+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Next
 
+* The is:haspower search works once again.
+* The is:cosmetic search will now work for Destiny 2.
+
 # 4.60.0 (2018-07-08)
 
 * Farming mode won't try to move unmoveable reputation tokens.