+++ docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Next
 
 * Fix a bug that would leave behind stackable items when moving certain loadouts like "Gather Reputation Items".
+* The is:haspower search works once again.
+* The is:cosmetic search will now work for Destiny 2.
 
 # 4.60.0 (2018-07-08)