+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
 
+# 4.61.0 (2018-07-15)
+
 * Fix a bug that would leave behind stackable items when moving certain loadouts like "Gather Reputation Items".
 * The is:haspower search works once again.
 * The is:cosmetic search will now work for Destiny 2.